third_party/blink/web_tests/external/wpt/focus/activeelement-after-calling-window-focus.sub-expected.txt

This is a testharness.js-based test.
[FAIL] Tests for all frames are in same origin
  assert_equals: expected "outerlog:windowblur,middlelog:innerlog:windowfocus,INPUT," but got "outerlog:middlelog:innerlog:INPUT,"
[FAIL] Tests for middle frame and inner frame are in same origin and outer frame is in different origin
  assert_equals: expected "outerlog:windowblur,middlelog:innerlog:windowfocus,INPUT," but got "outerlog:middlelog:innerlog:INPUT,"
[FAIL] Tests for outer frame and middle frame are in same origin and inner frame is in different origin
  assert_equals: expected "outerlog:windowblur,middlelog:innerlog:windowfocus,INPUT," but got "outerlog:middlelog:innerlog:INPUT,"
[FAIL] Tests for outer frame and inner frame are in same origin and middle frame is in different origin
  assert_equals: expected "outerlog:windowblur,middlelog:innerlog:windowfocus,INPUT," but got "outerlog:middlelog:innerlog:INPUT,"
[FAIL] Tests for all frames are in different origin
  assert_equals: expected "outerlog:windowblur,middlelog:innerlog:windowfocus,INPUT," but got "outerlog:middlelog:innerlog:INPUT,"
Harness: the test ran to completion.
